Rename StylableInlines.h to StyleableInlines.h rdar://173357919 https://bugs.webkit.org/show_bug.cgi?id=310737
Reviewed by Jer Noble.

Rename StylableInlines.h to StyleableInlines.h to match with the class name
(Styleable, Styleable.h). For those not seeing the difference, it's Styl*e*able.

Simple refactoring, no tests needed.

* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/animation/ScrollTimeline.cpp:
* Source/WebCore/animation/StyleOriginatedTimelinesController.cpp:
* Source/WebCore/animation/ViewTimeline.cpp:
* Source/WebCore/dom/Element.cpp:
* Source/WebCore/dom/PseudoElement.cpp:
* Source/WebCore/page/PageColorSampler.cpp:
* Source/WebCore/rendering/updating/RenderTreeUpdater.cpp:
* Source/WebCore/style/StyleAdjuster.cpp:
* Source/WebCore/style/Styleable.cpp:
* Source/WebCore/style/StyleableInlines.h: Renamed from Source/WebCore/style/StylableInlines.h.
(WebCore::Styleable::fromElement):

Canonical link: https://commits.webkit.org/309996@main
