+++ Source/WebCore/style/Styleable.cpp
@@ -48,7 +48,7 @@
 #include "RenderStyle+GettersInlines.h"
 #include "RenderView.h"
 #include "StyleAnimations.h"
-#include "StylableInlines.h"
+#include "StyleableInlines.h"
 #include "StyleChangedAnimatableProperties.h"
 #include "StyleCustomPropertyData.h"
 #include "StyleInterpolation.h"
